LPS-187069 use static inner class
